@@ -1,19 +1,15 @@
-require 'pry'
 class ClassificationsController < ApplicationController
   include ActionView::Helpers::TextHelper
   respond_to :json
 
   def create
+
     user = require_user!
 
     workflow_id      = BSON::ObjectId.from_string params["classifications"]["workflow_id"]
     task_key         = params["classifications"]["task_key"]
 
     annotation       = params["classifications"]["annotation"]
-    # if annotation.keys().include?("value")
-    #   annotation["value"] = invalid_line_ending_check(annotation["value"])
-    # end
-    
     annotation       = {} if annotation.nil?
     started_at       = params["classifications"]["metadata"]["started_at"]
     finished_at      = params["classifications"]["metadata"]["finished_at"]
@@ -31,13 +27,8 @@ class ClassificationsController < ApplicationController
       task_key: task_key,
       user: user
     )
-
-    respond_with @result
+    render json: @result
   end
-
-  # def invalid_line_ending_check(string)
-  #   string.gsub(/(?:\n\r?|\r\n?)/, '<br>')
-  # end
 
 
   def terms
