@@ -21,7 +21,7 @@ module API
       g.helper_specs false
     end
     
-    config.middleware.use Rack::Deflater
+    # config.middleware.use Rack::Deflater
 
     initializer 'setup_asset_pipeline', :group => :all  do |app|
       # We don't want the default of everything that isn't js or css, because it pulls too many things in
