@@ -1,7 +1,7 @@
 API::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
-  config.middleware.insert_before ActionDispatch::Static, Rack::Deflater
+  # config.middleware.insert_before ActionDispatch::Static, Rack::Deflater
   
   # Code is not reloaded between requests.
   config.cache_classes = true
